Convolutional AE/requirements.txt

audioread==2.1.5
h5py==2.10.0
Keras==2.1.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
librosa==0.7.2
matplotlib==3.0.3
numpy==1.16.0
plotly==4.6.0
PyYAML==5.1
scikit-learn==0.20.2
scipy==1.4.1
SoundFile==0.10.3.post1
spafe==0.1.1
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
tqdm==4.23.4

